Revert "Move creator function for sub-iterators to find-module"
This reverts commit fb9a6d67779e35a20b38fa3f841690918e5d7557.
